Remove fvs and rename backup_fvs to fields in show hamgrd actor command
### why
This is to address issue #78, which was raised during demo.

### what this PR does
Remove fvs and rename backup_fvs to fields in the internal table output in "show hamgrd actor" command. fvs is not removed from the message sent by hamgrd. It is only removed from the "view", which is output of the above command.