Implement mark-delete in actor framework (#104)
### why
actor has the retry logic in outgoing state. If a message is not acked,
it will resend the message to make sure receiver has received it
successfully. When an actor is terminated, the retry will be terminated
as well so it can't guarantee the receiver getting the message.

### what this PR does
Introduce mark-delete concept. 
1. when an actor is going to terminate, add "mark_deleted" flag to the
driver of the actor.
2. In the run loop of the actor, which is triggered each time it
receives a message, it will check if the actor is ready_for_delete.
3. ready_for_delete checks if there is unacked message in outgoing
state. Only exits from the run loop when there is non
4. When an actor is in mark_deleted state, stop processing incoming
requests but always replies OK. So 2 mark_deleted actors won't form a
dead loop.
5. response is processed normally so unacked messages can be ACKed.
6. management_request is processed normally so we can still dump actor
state using swbus-cli