Unregister handlers after actor terminates (#115)
### why
when actor terminates itself, the handler in SwbusEdgeRuntime is not
removed. When a new actor is spawned with the same service path, it will
be ignored because the ActorCreator replies on "NoRoute" to discover new
actor.

### what this PR does
when ActorDriver exits from run loop, the SimpleSwbusEdgeClient it owns
will be destructed. From the destructor, handler will be removed.
This addresses issue #111