Proper handling of entry not found in incoming vs decoding error (#114)
### why
currently Incoming::get returns error if the entry is not found and
caller typically propagates the error further. Sometimes it is normal
that an entry doesn't exist. It needs to be treated differently from
message decode error.

### what this PR does
Incoming::get returns Option. Caller needs to handle the None return
accordingly, which means the entry is not found.