Suppress route announcement if no change to advertised routes (#130)
<!--
Please make sure you have read and understood the contribution
guildlines:
https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

1. Make sure your commit includes a signature generted with `git commit
-s`
2. Make sure your commit title follows the correct format: [component]:
description
3. Make sure your commit message contains enough details about the
change and related tests
4. Make sure your pull request adds related reviewers, asignees, labels

Please also provide the following information in this pull request:
-->
**What I did**
If no new route is added, don't send route announcement. There might be
change to existing routes, such as path added or removed. But it won't
affect advertised routes because swbusd only advertises one route to its
neighbours no matter how many paths the route has.
**Why I did it**
This suppresses unnecessary route announcement. Current design will send
route announcement but the neighbour will not propagate to its neighbour
if there is no change to its route table. With the new design, it will
not send the announcement in the first place if there is no change.
**How I verified it**
Run it in a testbed with 2 switches, each with 8 swbusd. 
**Details if related**